Revert "coordis sql 모델 생성 (INSERT,SELECT)"
This reverts commit 98761fb5feb5f552ca0e5893597005bcb57cd39f.
